requirements.txt
discord.py==2.3.2
pillow>=10.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
asyncio>=3.4.3
python-dateutil>=2.8.2
